### Theory Fixed shift fancy magic bitboards are a variant of magic bitboards where instead of calculating a variable shift based on possible occupancy permutations of a given slide, a compile-constant shift is used. This the compiler can use less instructions to perform the shift by a constant, reducing indexing time to retrieve the attack bitboard.
Additionally it's possible to permute the pieces in the attack table in such a way that the end of square X attaks matches square Y attacks, given this it's possible to overlap entries, reducing the table size, _in theory_ improving cache locality. Not all magics are equal, different magics will result in different permutations of attacks of a certain square, and by continuing to test other magics it's possible to reduce the table size even more. It's unreasonable to compute this kind of a compressed table on engine startup, finding magics and their permutations which result in a small table takes hours and hours. Because of this, magics are pre-computed and hardcoded.

### Implementation
* Using magics and their offsets originally found by Volker Annuss and posted on http://www.talkchess.com/forum3/viewtopic.php?p=670709#p670709, the table size ends up being ~699 kB instead of 841kB.
* `RookTable` and `BishopTable` have been merged into one `SlideAttackTable` to allow overlapping.
* `Magic` class has been templated to allow constant shift, `index()` has been updated.
* Since the 32-bit implementations uses a different hashing mechanism, it's impossible to reuse the same magics for it and changing the hashing to be equivalent to the 64-bit version would have hurt performance more than the gain from fixed shift. This is the only reason why magics generation still remains, otherwise it could have been removed completely.

### Tests
I ran some tests on my machine on 64 bit with PEXT disabled. Test results are "hot" - meaning first few test results are dicsarded, this helps stabilize the results and decrease deviation. `bench 16 1 5 default perft`:

  | Fancy magics | Fixed shift fancy
-- | -- | --
  | 9938 ms | 9872 ms
  | 9946 ms | 9853 ms
  | 9914 ms | 9852 ms
  | 9924 ms | 9886 ms
  | 9911 ms | 9874 ms
  | 9929 ms | 9849 ms
  | 9930 ms | 9868 ms
  | 9910 ms | 9863 ms
  | 9913 ms | 9845 ms
  | 9914 ms | 9858 ms
  | 9934 ms | 9874 ms
  | 9917 ms | 9857 ms
  | 9932 ms | 9864 ms
  | 9933 ms | 9868 ms
  | 9912 ms | 9882 ms
  | 9926 ms | 9885 ms
  | 9927 ms | 9846 ms
  | 9910 ms | 9863 ms
  | 9941 ms | 9878 ms
  | 9928 ms | 9851 ms
Avg: | 9924 ms | 9864 ms
Std. dev | 11.18 ms | 12.91 ms

Improvements are quite small but consistent, average of **60 ms** or **0.61%**. PEXT and 32 bit performance seems unaffected.